Refactor Star and Solar System structures to new HATEOAS architecture.
