+++ api/src/main/java/com/senorpez/trident/api/StarController.java
@@ -1,7 +1,7 @@
 package com.senorpez.trident.api;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.hateoas.Resource;
+import org.springframework.hateoas.CollectionModel;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,12 +12,12 @@ import java.util.Collection;
 import java.util.stream.Collectors;
 
 import static com.senorpez.trident.api.SupportedMediaTypes.TRIDENT_API_VALUE;
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 @RequestMapping(
         value = "/systems/{solarSystemId}/stars",
         method = RequestMethod.GET,
-        produces = {TRIDENT_API_VALUE, APPLICATION_JSON_UTF8_VALUE}
+        produces = {TRIDENT_API_VALUE, APPLICATION_JSON_VALUE}
 )
 @RestController
 class StarController {
@@ -36,23 +36,23 @@ class StarController {
     }
 
     @RequestMapping
-    ResponseEntity<EmbeddedStarResources> stars(@PathVariable final int solarSystemId) {
+    ResponseEntity<CollectionModel<StarModel>> stars(@PathVariable final int solarSystemId) {
         final SolarSystem solarSystem = apiService.findOne(
                 this.solarSystems,
                 findSolarSystem -> findSolarSystem.getId() == solarSystemId,
                 () -> new SolarSystemNotFoundException(solarSystemId));
         final Collection<Star> stars = solarSystem.getStars();
-        final Collection<EmbeddedStarModel> starModels = stars.stream()
-                .map(EmbeddedStarModel::new)
-                .collect(Collectors.toList());
-        final Collection<Resource<EmbeddedStarModel>> starResources = starModels.stream()
-                .map(star -> star.toResource(solarSystemId))
-                .collect(Collectors.toList());
-        return ResponseEntity.ok(new EmbeddedStarResources(starResources, solarSystemId));
+        final CollectionModel<StarModel> starModels = CollectionModel.of(stars
+                .stream()
+                .map(EmbeddedStarEntity::new)
+                .map(embeddedStarEntity -> new StarModel(embeddedStarEntity, solarSystemId))
+                .collect(Collectors.toList())
+        );
+        return ResponseEntity.ok(starModels);
     }
 
     @RequestMapping("/{starId}")
-    ResponseEntity<StarResource> stars(@PathVariable final int solarSystemId, @PathVariable final int starId) {
+    ResponseEntity<StarModel> stars(@PathVariable final int solarSystemId, @PathVariable final int starId) {
         final SolarSystem solarSystem = apiService.findOne(
                 this.solarSystems,
                 findSolarSystem -> findSolarSystem.getId() == solarSystemId,
@@ -61,8 +61,8 @@ class StarController {
                 solarSystem.getStars(),
                 findStar -> findStar.getId() == starId,
                 () -> new StarNotFoundException(starId));
-        final StarModel starModel = new StarModel(star);
-        final StarResource starResource = starModel.toResource(solarSystemId);
-        return ResponseEntity.ok(starResource);
+        final StarEntity starEntity = new StarEntity(star);
+        final StarModel starModel = new StarModel(starEntity, solarSystemId);
+        return ResponseEntity.ok(starModel);
     }
 }
